Fixed a few more shadowed variables
* In AbstractDataStructureOberser, a parameter of the lambda shadows
a parameter of the function. For most cases, this wouldn't be an
issue unless the lambda captures the outer variable; however, the
name of the outer is still permitted to be used in the lambda in
certain expressions even when not captured. e.g. `sizeof`. Only gcc
warns on this case.

Signed-off-by: Jared Duffey <jared.duffey@bluequartz.net>