Fixing broken SetConditionalValue unit test.
The new macro was comparing 0.0 to the minimum float64 value, which is
BARELY above 0.0 and then it would fail.  Fixed this by creating a macro
that is used for direct conversions (for float32, float64, int64, uint64,
usize) and another separate macro used for conversions to a "container"
type which then gets casted down to the actual type.

Signed-off-by: Joey Kleingers <joey.kleingers@bluequartz.net>