CropImageGeometry: Fix runtime bugs where the proper data was not being copied.
The root of the problem was the flipping of the X and Z values when creating DataArrays
and Creating the Image Geometry.
Refactored many sections to be more efficient in performance and less code to maintain

Also fixed in this commit:
Fix issue in ParallelTaskAlgorithm if the developer set the parallelization to (false) nothing would be executed.
Add in some documentation to a few places to help developers understand the ordering of the axis.
Updated the test data file to include the exemplar data as SIMPL would produce it.

Signed-off-by: Michael Jackson <mike.jackson@bluequartz.net>