BUG: ReadCTFData-Remove phase=0 value adjustment. (#937)
There was a long-standing piece of code that added 1 to the phase value if the phase value was zero. This is not needed and can affect results being put out by DREAM3D.

Other changes in this PR were related to bringing consistency to the Parameter::Separator values for the 7 different filters to read EBSD data. Each of the filters should now present a consistent terminology to the user and be laid out the same.

Signed-off-by: Michael Jackson <mike.jackson@bluequartz.net>