BUG: Removes number of triangles guess from QuickMesh and SurfaceNets filters (#1233)
The filters were using a worst case scenario when attempting to create the triangle
geometry during preflight. This would lead to cases when the incoming ImageGeom has more
voxels than 2^31 which would cause preflight to fail. In reality the produced mesh will
most likely NOT have that many triangles.

The Write STL files filter will need to validate the number of triangles does
not exceed 2^31 triangles.

Signed-off-by: Michael Jackson <mike.jackson@bluequartz.net>