ENH: Clarify Kernel Average Misorientation algorithm with proper casting.
Added proper casting as the previous code was relying on the compiler to do the proper
casting at the proper time. This explicitly sets the types and adds defensive checks
to ensure the calculated index is >= 0.

Updated the documentation to state that the KAM is only calculated within a grain.

Signed-off-by: Michael Jackson <mike.jackson@bluequartz.net>