FILT: Create Vertex Geometry from Orientations (#1469)
This filter creates a vertex geometry from any set of Orientations. The orientations
can have any representation. The underlying orientations will be converted to a
quaternion that represents a positive rotation, then converted to a Stereographic 3D Coordinate which will serve as
the XYZ coordinate of the created vertex geometry.